refactor : 알람 설정한 고객이 있을 경우에만 메일 전송 기능과 알람 제거가 작동하도록 수정함

+++ src/main/java/com/wanted/preonboarding/ticketing/event/listener/CancelReservationEventListener.java
@@ -33,12 +33,18 @@ public class CancelReservationEventListener {
                 .findById(cancelReservationEvent.getPerformanceSeatInfoId())
                 .orElseThrow(() -> new NotFoundPerformanceSeatInfoException(ErrorCode.NOT_FOUND_PERFORMANCE_SEAT_INFO));
 
+        sendAlarm(performanceSeatInfo);
+    }
+
+    private void sendAlarm(PerformanceSeatInfo performanceSeatInfo) {
         List<Alarm> alarms = alarmRepository.findAllByPerformance(performanceSeatInfo.getPerformance());
 
-        for (Alarm alarm : alarms) {
-            emailSender.sendPerformanceInfo(alarm.getEmail(), performanceSeatInfo);
-        }
+        if (!alarms.isEmpty()) {
+            for (Alarm alarm : alarms) {
+                emailSender.sendPerformanceInfo(alarm.getEmail(), performanceSeatInfo);
+            }
 
-        alarmRepository.deleteAllInBatch(alarms);
+            alarmRepository.deleteAllInBatch(alarms);
+        }
     }
 }
